[REFACTOR/#220] Repository 인터페이스를 domain 패키지로 이동하고, RepositoryImpl을 단일표현식으로 통일합니다.
