+++ app/src/main/java/com/sopt/clody/data/repositoryimpl/DiaryRepositoryImpl.kt
@@ -8,7 +8,7 @@ import com.sopt.clody.data.remote.dto.response.MonthlyDiaryResponseDto
 import com.sopt.clody.data.remote.dto.response.ReplyDiaryResponseDto
 import com.sopt.clody.data.remote.dto.response.WriteDiaryResponseDto
 import com.sopt.clody.data.remote.util.handleApiResponse
-import com.sopt.clody.data.repository.DiaryRepository
+import com.sopt.clody.domain.repository.DiaryRepository
 import com.sopt.clody.presentation.utils.network.ErrorMessages
 import com.sopt.clody.presentation.utils.network.ErrorMessages.FAILURE_TEMPORARY_MESSAGE
 import retrofit2.HttpException
@@ -17,20 +17,18 @@ import javax.inject.Inject
 class DiaryRepositoryImpl @Inject constructor(
     private val diaryRemoteDataSource: DiaryRemoteDataSource
 ): DiaryRepository {
-    override suspend fun writeDiary(date: String, content: List<String>): Result<WriteDiaryResponseDto> {
-        return runCatching {
+    override suspend fun writeDiary(date: String, content: List<String>): Result<WriteDiaryResponseDto> =
+        runCatching {
             diaryRemoteDataSource.writeDiary(date, content).handleApiResponse().getOrThrow()
         }
-    }
 
-    override suspend fun deleteDailyDiary(year: Int, month: Int, day: Int): Result<DailyDiariesResponseDto> {
-        return runCatching {
+    override suspend fun deleteDailyDiary(year: Int, month: Int, day: Int): Result<DailyDiariesResponseDto> =
+        runCatching {
             diaryRemoteDataSource.deleteDailyDiary(year, month, day).handleApiResponse().getOrThrow()
         }
-    }
 
-    override suspend fun getDailyDiariesData(year: Int, month: Int, date: Int): Result<DailyDiariesResponseDto> {
-        return runCatching {
+    override suspend fun getDailyDiariesData(year: Int, month: Int, date: Int): Result<DailyDiariesResponseDto> =
+        runCatching {
             diaryRemoteDataSource.getDailyDiariesData(year, month, date).handleApiResponse()
         }.getOrElse { exception ->
             val errorMessage = when (exception) {
@@ -45,16 +43,14 @@ class DiaryRepositoryImpl @Inject constructor(
             }
             Result.failure(Exception(errorMessage))
         }
-    }
 
-    override suspend fun getDiaryTime(year: Int, month: Int, date: Int): Result<DiaryTimeResponseDto> {
-        return runCatching {
+    override suspend fun getDiaryTime(year: Int, month: Int, date: Int): Result<DiaryTimeResponseDto> =
+        runCatching {
             diaryRemoteDataSource.getDiaryTime(year, month, date).data
         }
-    }
 
-    override suspend fun getMonthlyCalendarData(year: Int, month: Int): Result<MonthlyCalendarResponseDto> {
-        return runCatching {
+    override suspend fun getMonthlyCalendarData(year: Int, month: Int): Result<MonthlyCalendarResponseDto> =
+        runCatching {
             diaryRemoteDataSource.getMonthlyCalendarData(year, month).handleApiResponse()
         }.getOrElse { exception ->
             val errorMessage = when (exception) {
@@ -70,21 +66,18 @@ class DiaryRepositoryImpl @Inject constructor(
             }
             Result.failure(Exception(errorMessage))
         }
-    }
 
-    override suspend fun getMonthlyDiary(year: Int, month: Int): Result<MonthlyDiaryResponseDto> {
-        return runCatching {
+    override suspend fun getMonthlyDiary(year: Int, month: Int): Result<MonthlyDiaryResponseDto> =
+        runCatching {
             diaryRemoteDataSource.getMonthlyDiary(year, month).handleApiResponse().getOrThrow()
         }
-    }
 
-    override suspend fun getReplyDiary(year: Int, month: Int, date: Int): Result<ReplyDiaryResponseDto> {
-        return runCatching {
+    override suspend fun getReplyDiary(year: Int, month: Int, date: Int): Result<ReplyDiaryResponseDto> =
+        runCatching {
             val response = diaryRemoteDataSource.getReplyDiary(year, month, date).data
             if (response.content == null) {
                 throw IllegalStateException(FAILURE_TEMPORARY_MESSAGE)
             }
             response
         }
-    }
 }
