+++ app/src/main/java/com/sopt/clody/presentation/ui/auth/signup/SignUpViewModel.kt
@@ -8,8 +8,8 @@ import com.kakao.sdk.user.UserApiClient
 import com.sopt.clody.ClodyFirebaseMessagingService
 import com.sopt.clody.data.remote.dto.request.LoginRequestDto
 import com.sopt.clody.data.remote.dto.request.SignUpRequestDto
-import com.sopt.clody.data.repository.AuthRepository
-import com.sopt.clody.data.repository.TokenRepository
+import com.sopt.clody.domain.repository.AuthRepository
+import com.sopt.clody.domain.repository.TokenRepository
 import com.sopt.clody.presentation.utils.base.UiState
 import com.sopt.clody.presentation.utils.network.ErrorMessages.FAILURE_NETWORK_MESSAGE
 import com.sopt.clody.presentation.utils.network.ErrorMessages.FAILURE_TEMPORARY_MESSAGE
